requirements.txt
aiogram
opencv-python
numpy
pillow
easyocr
ultralytics
python-dotenv
openai
fastapi
uvicorn
sqlalchemy
aiosqlite
pydantic-settings
